requirements.txt
python-dotenv==1.0.1
pymongo==4.10.1
pandas==2.2.3